refactor: reorganize distroless oci packages
