+++ bazel/cmd/knife/cmd/debian/versions.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/arkeros/senku/distroless/pkg/lockfile"
+	"github.com/arkeros/senku/oci/distroless/debian/lockfile"
 )
 
 type versionsOptions struct {
@@ -23,8 +23,8 @@ func newCmdVersions() *cobra.Command {
 		Long: `Extracts and displays package name and version pairs from a Debian lock file.
 
 Examples:
-  knife debian versions distroless/debian13.lock.json
-  knife debian versions --arch amd64 distroless/debian13.lock.json`,
+  knife debian versions oci/distroless/debian13.lock.json
+  knife debian versions --arch amd64 oci/distroless/debian13.lock.json`,
 		Args: cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			o.Path = args[0]
