+++ oci/BUILD
@@ -2,7 +2,7 @@ load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
 
 bzl_library(
     name = "oci",
-    srcs = ["oci.bzl"],
+    srcs = ["oci_image.bzl"],
     visibility = ["//visibility:public"],
     deps = [
         "@grype.bzl//:grype",
@@ -23,7 +23,7 @@ bzl_library(
     deps = [
         ":config",
         ":oci",
-        "//distroless/common:variables",
+        "//oci/distroless/common:variables",
         "@bazel_lib//lib:transitions",
         "@rules_img//img:image",
     ],
@@ -36,8 +36,8 @@ bzl_library(
     deps = [
         ":config",
         ":oci",
-        "//distroless:distro",
-        "//distroless/common:variables",
+        "//oci/distroless:distro",
+        "//oci/distroless/common:variables",
         "@aspect_rules_js//js:defs",
         "@aspect_rules_py//py:defs",
         "@container_structure_test//:defs",
@@ -52,10 +52,23 @@ bzl_library(
     deps = [
         ":config",
         ":oci",
-        "//distroless:distro",
-        "//distroless/common:variables",
+        "//oci/distroless:distro",
+        "//oci/distroless/common:variables",
         "@aspect_rules_py//py:defs",
         "@container_structure_test//:defs",
         "@rules_img//img:image",
+    ],
+)
+
+bzl_library(
+    name = "oci_image",
+    srcs = ["oci_image.bzl"],
+    visibility = ["//visibility:public"],
+    deps = [
+        "@grype",
+        "@rules_img//img:image",
+        "@rules_img//img:load",
+        "@rules_img//img:push",
+        "@tar",
     ],
 )
