+++ oci/distroless/BUILD
@@ -4,4 +4,9 @@ bzl_library(
     name = "distro",
     srcs = ["distro.bzl"],
     visibility = ["//visibility:public"],
+    deps = [
+        "//oci",
+        "//oci/distroless/common:variables",
+        "@rules_img//img:image",
+    ],
 )
