+++ oci/distroless/cc/BUILD
@@ -1,7 +1,8 @@
 load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
 load("@container_structure_test//:defs.bzl", "container_structure_test")
 load("@rules_distroless//distroless:defs.bzl", "flatten")
-load(":base.bzl", "base_image", "base_image_index")
+load("//oci/distroless:distro.bzl", "distroless_matrix")
+load("//oci/distroless/common:variables.bzl", "COMPRESSION")
 load(":config.bzl", "BASE_ARCHITECTURES", "BASE_DISTROS", "BASE_PACKAGES")
 
 package(default_visibility = ["//visibility:public"])
@@ -9,6 +10,7 @@ package(default_visibility = ["//visibility:public"])
 [
     flatten(
         name = arch + "_" + distro + "_layer",
+        compress = COMPRESSION,
         deduplicate = True,
         tars = [
             "@{}//{}/{}".format(distro, pkg, arch)
@@ -20,18 +22,15 @@ package(default_visibility = ["//visibility:public"])
 ]
 
 [
-    base_image(
-        arch = arch,
-        distro = distro,
-    )
-    for distro in BASE_DISTROS
-    for arch in BASE_ARCHITECTURES[distro]
-]
-
-[
-    base_image_index(
+    distroless_matrix(
+        name = "base",
         architectures = BASE_ARCHITECTURES[distro],
+        base = "//oci/distroless/static:static{mode}_{user}_{arch}_{distro}",
         distro = distro,
+        layers = [
+            ":{arch}_{distro}_layer",
+            "@{distro}//libstdc++6/{arch}",
+        ],
     )
     for distro in BASE_DISTROS
 ]
@@ -43,16 +42,6 @@ container_structure_test(
     driver = "docker",
     image = ":base_nonroot_amd64_debian13.oci_layout",
     tags = ["requires-docker"],
-)
-
-bzl_library(
-    name = "base",
-    srcs = ["base.bzl"],
-    deps = [
-        "//distroless/common:variables",
-        "//oci",
-        "@rules_img//img:image",
-    ],
 )
 
 bzl_library(
