+++ oci/distroless/common/variables.bzl
