+++ oci/distroless/debian/lockfile/testdata/cross_arch.lock.json
