+++ oci/distroless/debian/snapshot/testdata/manifest.yaml
