+++ oci/distroless/debian/snapshot/testdata/manifest_with_header.yaml
