+++ oci/distroless/nginx-mainline.yaml
