+++ oci/distroless/nginx-stable.lock.json
