+++ oci/distroless/nginx-stable.yaml
