+++ oci/distroless/nginx/config.bzl
