+++ oci/distroless/static/config.bzl
