+++ oci/js_image.bzl
@@ -2,9 +2,9 @@ load("@aspect_rules_js//js:defs.bzl", "js_binary", "js_image_layer")
 load("@aspect_rules_py//py:defs.bzl", "py_image_layer")
 load("@container_structure_test//:defs.bzl", "container_structure_test")
 load("@rules_img//img:image.bzl", "image_index")
-load("//distroless:distro.bzl", "ARCHITECTURE_PLATFORMS")
-load("//distroless/common:variables.bzl", "COMPRESSION", "DEBUG_MODE", "USERS", "USER_IDS")
-load(":oci.bzl", "oci_image")
+load("//oci/distroless:distro.bzl", "ARCHITECTURE_PLATFORMS")
+load("//oci/distroless/common:variables.bzl", "COMPRESSION", "DEBUG_MODE", "USERS", "USER_IDS")
+load(":oci_image.bzl", "oci_image")
 load(":config.bzl", "NODEJS_ARCHITECTURES", "NODEJS_DISTROS")
 
 def js_image_index(name, distro, architectures):
@@ -49,7 +49,7 @@ def js_image(
     [
         oci_image(
             name = name + mode + "_" + user + "_" + arch,
-            base = "//distroless/bash:bash" + mode + "_" + user + "_" + arch + "_" + distro,
+            base = "//oci/distroless/bash:bash" + mode + "_" + user + "_" + arch + "_" + distro,
             entrypoint = entrypoint,
             working_dir = "/",
             env = {
